+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`3.0.0...main`][3.0.0...main].
 ### Changed
 
 - Required `ergebnis/json-pointer:^3.0.0` and `ergebnis/json-schema-validator:^3.1.0` ([#697]), by [@localheinz]
+- Required `ergebnis/json-pointer:^3.1.0` ([#698]), by [@dependabot]
 
 ## [`3.0.0`][3.0.0]
 
@@ -502,6 +503,7 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [#672]: https://github.com/ergebnis/json-normalizer/pull/672
 [#673]: https://github.com/ergebnis/json-normalizer/pull/673
 [#697]: https://github.com/ergebnis/json-normalizer/pull/697
+[#698]: https://github.com/ergebnis/json-normalizer/pull/698
 
 [@BackEndTea]: https://github.com/BackEndTea
 [@dependabot]: https://github.com/dependabot