+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`3.0.0...main`][3.0.0...main].
+For a full diff see [`3.0.1...main`][3.0.1...main].
+
+## [`3.0.1`][3.0.1]
+
+For a full diff see [`3.0.0...3.0.1`][3.0.0...3.0.1].
+
+### Fixed
+
+- Adjusted `ConfigHashNormalizer` to sort keys correctly ([#723]), by [@fredded]
 
 ### Changed
 
@@ -385,6 +393,7 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [2.1.0]: https://github.com/ergebnis/json-normalizer/releases/tag/2.1.0
 [2.2.0]: https://github.com/ergebnis/json-normalizer/releases/tag/2.2.0
 [3.0.0]: https://github.com/ergebnis/json-normalizer/releases/tag/3.0.0
+[3.0.1]: https://github.com/ergebnis/json-normalizer/releases/tag/3.0.1
 
 [5d8b3e2...0.1.0]: https://github.com/ergebnis/json-normalizer/compare/5d8b3e2...0.1.0
 [0.1.0...0.2.0]: https://github.com/ergebnis/json-normalizer/compare/0.1.0...0.2.0
@@ -413,7 +422,8 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [2.0.0...2.1.0]: https://github.com/ergebnis/json-normalizer/compare/2.0.0...2.1.0
 [2.1.0...2.2.0]: https://github.com/ergebnis/json-normalizer/compare/2.1.0...2.2.0
 [2.2.0...3.0.0]: https://github.com/ergebnis/json-normalizer/compare/2.2.0...3.0.0
-[3.0.0...main]: https://github.com/ergebnis/json-normalizer/compare/3.0.0...main
+[3.0.0...3.0.1]: https://github.com/ergebnis/json-normalizer/compare/3.0.0...3.0.1
+[3.0.1...main]: https://github.com/ergebnis/json-normalizer/compare/3.0.1...main
 
 [#1]: https://github.com/ergebnis/json-normalizer/pull/1
 [#2]: https://github.com/ergebnis/json-normalizer/pull/2
@@ -504,8 +514,10 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [#673]: https://github.com/ergebnis/json-normalizer/pull/673
 [#697]: https://github.com/ergebnis/json-normalizer/pull/697
 [#698]: https://github.com/ergebnis/json-normalizer/pull/698
+[#723]: https://github.com/ergebnis/json-normalizer/pull/723
 
 [@BackEndTea]: https://github.com/BackEndTea
 [@dependabot]: https://github.com/dependabot
 [@ergebnis]: https://github.com/ergebnis
+[@fredden]: https://github.com/fredden
 [@localheinz]: https://github.com/localheinz